Add script to download and parse Project Gutenberg books
This commit introduces a new script `download-gutenberg.py` to automate the download and parsing of books from Project Gutenberg. Additionally, it includes sample RDF and YAML metadata files for "The Complete Works of William Shakespeare" and updates the `.gitignore` file to exclude certain directories. Dependencies in `requirements.txt` have been changed to use greater-than-or-equal-to (>=) versions.